requirements.txt
python-dotenv
pandas
pandas_ta
numpy
pymongo
certifi
requests


##-e .